added toggle-check-box-mixin
